Use clock framework rate for I2C driver
This removes the default hard-coded frequency setting for the I2C
driver and moves this setting to the clock framework instead,
where it is defined to 24MHz for the ED ASIC and 48MHz for the V1.
The default frequency #define has been removed to avoid confusion,
since this frequency is not defined by the I2C block but by the
clocking infrastructure.

Signed-off-by: Linus Walleij <linus.walleij@stericsson.com>