ux500-gpio: replace GPIO_ALTF_* macros with NMK_GPIO_*
The main reason for this change is that the GPIO_ALTF_* macros
were supposed to be passed to gpio_set_value in the old driver,
but this is not supported anymore.  We remove the old macros
so that there is no silent breakage in any out-of-tree
drivers.  NMK_GPIO_ALT_* macros are to be passed to the
nomadik_gpio_set_mode() function.

Signed-off-by: Rabin Vincent <rabin.vincent@stericsson.com>