HCL driver implementation (hash).
Major refactoring, checkpatch fixes, documentation of functions,
structures, enums. Also moved a lot of code from hash_alg.c to hash_core.c.

Fixed problem with incorrect digest when doing HMAC calculations. The
problem was that the function that did message pad seems to handle padding
for keys incorrectly. Now we do not say how many valid bits there are in
the last word when it comes to the key and then the hardware seems to
handle the key correctly.

Fixed contextsaving so hmac(sha1) passes.
There was an if statement checking the DINF bit which was never set. I
have removed this and then contextsaving for hmac(sha1) is working.

Code up to date with the new arch/arm/mach-ux500 folder structure.

Signed-off-by: Joakim Bech <joakim.xx.bech@stericsson.com>