B2R2: Changed behaviour of log prints.
The use of printk for error and debug prints has been changed to a set
of log macros. The verbosity of the log prints can now be controlled
through debugfs.

If the kernel is compiled with support for dynamic
debug (Documentation/dynamic-debug-howto.txt), this will be used
instead of the debugfs files.

ST-Ericsson Change ID: ER261401

Signed-off-by: Fredrik Allansson <fredrik.allansson@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/502
Reviewed-by: Robert LIND <robert.lind@stericsson.com>
Reviewed-by: Robert FEKETE <robert.fekete@stericsson.com>
Tested-by: Robert FEKETE <robert.fekete@stericsson.com>
Signed-off-by: Mian Yousaf Kaukab <mian.yousaf.kaukab@stericsson.com>
Change-Id: Ifbbc09525636ef396d4814ce21b61861983fbbe0
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/2377
Reviewed-by: Jonas ABERG <jonas.aberg@stericsson.com>