leds-ab8500: support backlight feature for secondary lcd
pwm1 is being used for primary lcd and pwm2 is used by sencondary
lcd. checks if secondary lcd is enabled and if so monitors the
backlight(intensity) of secondary lcd.
LCD backlight control in android standard application is supported
for only primary lcd, the same is replicated for secondary lcd.
Only one interface is provided by name "lcd-backlight" in sysfs,
which controls both the lcd's.

ST-Ericsson ID: ER262317

Signed-off-by: Arun Murthy <arun.murthy@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/556
Reviewed-by: Jonas ABERG <jonas.aberg@stericsson.com>
Tested-by: Jonas ABERG <jonas.aberg@stericsson.com>
Signed-off-by: Mian Yousaf Kaukab <mian.yousaf.kaukab@stericsson.com>
Change-Id: I989d67eeea75730f9b863cbe4f29c3c1da3305e0
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/2387