Bluetooth antenna problems on Hrefp boards
This fix activates the PTA output of the BT chip, which will
trigger the PTA circuit in the Wlan chip so that BT can share
the same antenna as Wlan and function properly. Also this fix
changes from space to tab in the 2 info files. This is because
the factory that flash the firmware will have problems otherwise.

ST-Ericsson ID: ER266093

Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/2703
Reviewed-by: Par-Gunnar HJALMDAHL <par-gunnar.p.hjalmdahl@stericsson.com>
Reviewed-by: Andreas ARESKOUG <andreas.areskoug@stericsson.com>
Tested-by: Andreas ARESKOUG <andreas.areskoug@stericsson.com>
Reviewed-by: Jonas ABERG <jonas.aberg@stericsson.com>
Signed-off-by: Mian Yousaf Kaukab <mian.yousaf.kaukab@stericsson.com>
Change-Id: I5a92416b2269d03ad51cf2899cd77a3e160a9396
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/2844